Presentation- Briefing about the "Smart Car"
My team and I had designed a Smart Car that could automatically slow down/stop when the car gets closer than a specified distance to another car/person/object in front. This is done to prevent accidents when the drivers are careless or distracted to notice the obstacle and slow down the speed appropriately. As soon as the obstacle is removed/goes further away, the car gradually picks up its original speed. When this automatic control happens, the driver is notified by a buzzer sound. The driver may also choose to stop this automatic speed control anytime.